+++ minijinja/src/source.rs
@@ -1,14 +1,17 @@
-use std::collections::BTreeMap;
+use std::collections::HashMap;
 use std::fmt;
 use std::fs;
 use std::path::Path;
+use std::sync::Arc;
 
+use memo_map::MemoMap;
 use self_cell::self_cell;
 
 use crate::environment::CompiledTemplate;
 use crate::error::{Error, ErrorKind};
-use crate::utils::BTreeMapKeysDebug;
 use crate::value::RcType;
+
+type LoadFunc = dyn for<'a> Fn(&'a str) -> Result<String, Error> + Send + Sync;
 
 /// Utility for dynamic template loading.
 ///
@@ -21,7 +24,11 @@ use crate::value::RcType;
 /// [`Environment::add_template`](crate::Environment::add_template) must no
 /// longer be used as otherwise the same lifetime concern arises.
 ///
-/// # Example
+/// Alternatively sources can also be used to implement completely dynamic template
+/// lookups by using [`with_loader`](Source::with_loader) in which case templates
+/// are loaded on first use.
+///
+/// # Static Example
 ///
 /// ```rust
 /// # use minijinja::{Source, Environment};
@@ -33,15 +40,58 @@ use crate::value::RcType;
 ///     env
 /// }
 /// ```
-#[derive(Clone, Default)]
+///
+/// # Fully Dynamic Example
+///
+/// ```rust
+/// # use minijinja::{Source, Environment};
+/// fn create_env() -> Environment<'static> {
+///     let mut env = Environment::new();
+///     env.set_source(Source::with_loader(|name| {
+///         if name == "layout.html" {
+///             Ok(Some("...".into()))
+///         } else {
+///             Ok(None)
+///         }
+///     }));
+///     env
+/// }
+/// ```
+#[derive(Clone)]
 #[cfg_attr(docsrs, doc(cfg(feature = "source")))]
 pub struct Source {
-    templates: BTreeMap<String, RcType<LoadedTemplate>>,
+    backing: SourceBacking,
+}
+
+#[derive(Clone)]
+enum SourceBacking {
+    Dynamic {
+        templates: MemoMap<String, RcType<LoadedTemplate>>,
+        loader: Arc<LoadFunc>,
+    },
+    Static {
+        templates: HashMap<String, RcType<LoadedTemplate>>,
+    },
+}
+
+impl Default for Source {
+    fn default() -> Source {
+        Source::new()
+    }
 }
 
 impl fmt::Debug for Source {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt::Debug::fmt(&BTreeMapKeysDebug(&self.templates), f)
+        match &self.backing {
+            SourceBacking::Dynamic { templates, .. } => f
+                .debug_list()
+                .entries(templates.iter().map(|x| x.0))
+                .finish(),
+            SourceBacking::Static { templates } => f
+                .debug_list()
+                .entries(templates.iter().map(|x| x.0))
+                .finish(),
+        }
     }
 }
 
@@ -62,7 +112,32 @@ impl fmt::Debug for LoadedTemplate {
 impl Source {
     /// Creates an empty source.
     pub fn new() -> Source {
-        Source::default()
+        Source {
+            backing: SourceBacking::Static {
+                templates: HashMap::new(),
+            },
+        }
+    }
+
+    /// Sets a dynamic loader function.
+    ///
+    /// When a loader is set the source gains the ability to dynamically
+    /// load templates.  The loader is invoked with the name of the template.
+    /// If this template exists `Ok(Some(template_source))` has to be returned,
+    /// otherwise `Ok(None)`.  It's also possible to signal out other errors.
+    pub fn with_loader<F>(f: F) -> Source
+    where
+        F: Fn(&str) -> Result<Option<String>, Error> + Send + Sync + 'static,
+    {
+        Source {
+            backing: SourceBacking::Dynamic {
+                templates: MemoMap::new(),
+                loader: Arc::new(move |name| match f(name)? {
+                    Some(rv) => Ok(rv),
+                    None => Err(Error::new_not_found(name)),
+                }),
+            },
+        }
     }
 
     /// Adds a new template into the source.
@@ -81,13 +156,28 @@ impl Source {
         let tmpl = LoadedTemplate::try_new(owner, |(name, source)| -> Result<_, Error> {
             CompiledTemplate::from_name_and_source(name.as_str(), source)
         })?;
-        self.templates.insert(name, RcType::new(tmpl));
+
+        match self.backing {
+            SourceBacking::Dynamic { ref templates, .. } => {
+                templates.insert(name, RcType::new(tmpl));
+            }
+            SourceBacking::Static { ref mut templates } => {
+                templates.insert(name, RcType::new(tmpl));
+            }
+        }
         Ok(())
     }
 
     /// Removes an already loaded template from the source.
     pub fn remove_template(&mut self, name: &str) {
-        self.templates.remove(name);
+        match &mut self.backing {
+            SourceBacking::Dynamic { templates, .. } => {
+                templates.remove(name);
+            }
+            SourceBacking::Static { templates } => {
+                templates.remove(name);
+            }
+        }
     }
 
     /// Loads templates from a path.
@@ -159,9 +249,23 @@ impl Source {
     }
 
     /// Gets a compiled template from the source.
-    pub(crate) fn get_compiled_template(&self, name: &str) -> Option<&CompiledTemplate<'_>> {
-        self.templates
-            .get(name)
-            .map(|value| value.borrow_dependent())
+    pub(crate) fn get_compiled_template(&self, name: &str) -> Result<&CompiledTemplate<'_>, Error> {
+        match &self.backing {
+            SourceBacking::Dynamic { templates, loader } => Ok(templates
+                .get_or_try_insert(name, || -> Result<_, Error> {
+                    let source = loader(name)?;
+                    let owner = (name.to_owned(), source);
+                    let tmpl =
+                        LoadedTemplate::try_new(owner, |(name, source)| -> Result<_, Error> {
+                            CompiledTemplate::from_name_and_source(name.as_str(), source)
+                        })?;
+                    Ok(RcType::new(tmpl))
+                })?
+                .borrow_dependent()),
+            SourceBacking::Static { templates } => templates
+                .get(name)
+                .map(|value| value.borrow_dependent())
+                .ok_or_else(|| Error::new_not_found(name)),
+        }
     }
 }